add editable

+++ js/script.js
@@ -25,7 +25,8 @@ class ToDo {
     li.key = todo.key;
     li.insertAdjacentHTML('beforeend', `
         <span class="text-todo">${todo.value}</span>
-				<div class="todo-buttons">
+        <div class="todo-buttons">
+          <button class="todo-edit"></button>
 					<button class="todo-remove"></button>
 					<button class="todo-complete"></button>
 				</div>
@@ -56,6 +57,24 @@ class ToDo {
   generateKey() {
     return Math.random().toString(36).substring(2, 15) + Math.random().toString(36).substring(2, 15);
   }
+  // редактировать
+  editItem(editText, editBtn) {
+    editText.contentEditable = true;
+    editText.tabindex = 0;
+    editText.focus();
+    editBtn.style.display = 'none';
+  }
+  // сохранить изменения после редактирования
+  saveChangings(editText, key, editBtn) {
+    editText.contentEditable = false;
+    this.todoData.get(key).value = editText.textContent;
+    editBtn.style.display = '';
+    this.render();
+    editText.removeEventListener('blur', () => {
+      this.saveChangings(editText, key, editBtn);
+    });
+  }
+
   // удалить
   deleteItem(key) {
     this.todoData.delete(key);
@@ -70,13 +89,23 @@ class ToDo {
   handler() {
     this.todoContainer.addEventListener('click', event => {
       const target = event.target;
-      const  currentKey = target.closest('li').key;
-      if (target.matches('.todo-complete')) {
-        this.completedItem(currentKey);
-      }
-      if (target.matches('.todo-remove')) {
-        target.closest('li').remove();
-        this.deleteItem(currentKey);
+      if (target.closest('li')) {
+        const  currentKey = target.closest('li').key;
+        if (target.matches('.todo-complete')) {
+          this.completedItem(currentKey);
+        }
+        if (target.matches('.todo-remove')) {
+          target.closest('li').remove();
+          this.deleteItem(currentKey);
+        }
+        if (target.matches('.todo-edit')) {
+          const editText = target.closest('li').querySelector('.text-todo');
+          const editBtn = target.closest('li').querySelector('.todo-edit');
+          this.editItem(editText, editBtn);
+          editText.addEventListener('blur', () => {
+            this.saveChangings(editText, currentKey, editBtn);
+          });
+        }
       }
     });
   }
